requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.4
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.0.1
pillow-heif==0.13.0
requests==2.31.0
tqdm==4.66.1
urllib3==1.26.6
